+++ libraries/resource_providers/api500/c7000/storage_pool_provider.rb
@@ -61,6 +61,7 @@ module OneviewCookbook
         def update
           load_storage_system
           desired_state = Marshal.load(Marshal.dump(@item.data))
+	  desired_state.delete('type')
           @item.retrieve! || raise("Resource not found: The #{@resource_name} '#{@name}' could not be found")
           update_manage_state(desired_state['isManaged']) unless desired_state['isManaged'].nil?
           return Chef::Log.info("#{@resource_name} '#{@name}' has no need to update") if @item.like? desired_state