+++ examples/uplink_set.rb
@@ -1,4 +1,4 @@
-# (c) Copyright 2016 Hewlett Packard Enterprise Development LP
+# (c) Copyright 2020 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -20,6 +20,7 @@ my_client = {
   url: ENV['ONEVIEWSDK_URL'],
   user: ENV['ONEVIEWSDK_USER'],
   password: ENV['ONEVIEWSDK_PASSWORD']
+  api_version: 1600
 }
 
 # Example: creates or updates an uplink set
@@ -40,6 +41,7 @@ oneview_uplink_set 'UplinkSet1' do
     description: 'Created by Chef',
     portConfigInfos:
     [
+      desiredSpeed: 'Auto',	    
       location:
       {
         locationEntries: