+++ examples/fc_network.rb
@@ -16,7 +16,7 @@ my_client = {
   url: ENV['ONEVIEWSDK_URL'],
   user: ENV['ONEVIEWSDK_USER'],
   password: ENV['ONEVIEWSDK_PASSWORD'],
-  api_version: 1200
+  api_version: 1600
 }
 
 # Example: Create and manage a new fc network