+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - This release extends support of the SDK to OneView REST API version 1600 (OneView v5.20).
 
 ### Oneview Features supported
+- oneview_logical_interconnect
 - oneview_server_hardware
 - oneview_server_hardware_type