+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - oneview_hypervisor_manager
 - oneview_interconnect
 - oneview_logical_interconnect
+- oneview_scope
 - oneview_server_certificate
 - oneview_server_hardware
 - oneview_server_hardware_type