+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Oneview Features supported
 - oneview_hypervisor_manager
 - oneview_interconnect
+- oneview_logical_enclosure
 - oneview_logical_interconnect
 - oneview_server_certificate
 - oneview_server_hardware