+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - oneview_server_certificate
 - oneview_server_hardware
 - oneview_server_hardware_type
+- oneview_server_profile
 - oneview_server_profile_template
 
 ## 3.3.0