+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 3.4.0(unreleased)
 ### Notes
-- This release extends support of the SDK to OneView REST API version 1600 (OneView v5.20).
+- This release extends support of the SDK to OneView REST API version 1600 (OneView v5.20) and (ImageStreamer v5.20).
 
 ### Oneview Features supported
+- image_streamer_artifact_bundle
 - oneview_enclosure
 - oneview_hypervisor_manager
 - oneview_interconnect