+++ CHANGELOG.md
@@ -16,6 +16,9 @@
 - oneview_server_profile
 - oneview_server_profile_template
 
+Adds API1600 support to HPE Synergy Image Streamer resources:
+  - image_streamer_deployment_plan
+
 ## 3.3.0
 ### Notes
 - This release adds support to Oneview Rest API version 800, 1000, and 1200 for C7000 and Synergy to the features Hypervisor Manager and Hypervisor Cluster Profile, and Oneview Rest API version 600, 800, 1000, and 1200 for C7000 and Synergy to the feature Server Certificate.