+++ CHANGELOG.md
@@ -16,6 +16,10 @@
 - oneview_server_hardware_type
 - oneview_server_profile
 - oneview_server_profile_template
+- oneview_storage_pool
+- oneview_storage_system
+- oneview_storage_volume
+- oneview_volume_template
 
 Adds API1600 support to HPE Synergy Image Streamer resources:
   - image_streamer_deployment_plan