+++ CHANGELOG.md
@@ -3,11 +3,11 @@
 - This release extends support of the SDK to OneView REST API version 1600 (OneView v5.20) and (ImageStreamer v5.20).
 
 ### Oneview Features supported
+- image_streamer_artifact_bundle
 - oneview_connection_template
 - oneview_ethernet_network
 - oneview_fc_network
 - oneview_fcoe_network
-- image_streamer_artifact_bundle
 - oneview_enclosure
 - oneview_hypervisor_cluster_profile
 - oneview_hypervisor_manager
@@ -15,12 +15,17 @@
 - oneview_logical_enclosure
 - oneview_logical_interconnect
 - oneview_network_set
+- oneview_scope
 - oneview_server_certificate
 - oneview_server_hardware
 - oneview_server_hardware_type
 - oneview_uplink_set
 - oneview_server_profile
 - oneview_server_profile_template
+- oneview_storage_pool
+- oneview_storage_system
+- oneview_storage_volume
+- oneview_volume_template
 
 Adds API1600 support to HPE Synergy Image Streamer resources:
   - image_streamer_deployment_plan