+++ CHANGELOG.md
@@ -4,17 +4,32 @@
 
 ### Oneview Features supported
 - image_streamer_artifact_bundle
+- oneview_connection_template
+- oneview_enclosure_group
+- oneview_ethernet_network
+- oneview_fc_network
+- oneview_fcoe_network
 - oneview_enclosure
 - oneview_hypervisor_cluster_profile
 - oneview_hypervisor_manager
 - oneview_interconnect
 - oneview_logical_enclosure
 - oneview_logical_interconnect
+- oneview_network_set
+- oneview_scope
 - oneview_server_certificate
 - oneview_server_hardware
 - oneview_server_hardware_type
 - oneview_server_profile
 - oneview_server_profile_template
+- oneview_storage_pool
+- oneview_storage_system
+- oneview_storage_volume
+- oneview_volume_template
+- oneview_uplink_set
+
+Adds API1600 support to HPE Synergy Image Streamer resources:
+  - image_streamer_deployment_plan
 
 ## 3.3.0
 ### Notes