+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - oneview_interconnect
 - oneview_logical_enclosure
 - oneview_logical_interconnect
+- oneview_logical_interconnect_group
 - oneview_network_set
 - oneview_scope
 - oneview_server_certificate